Allow fractional rate in spnl, for percolation to run less than every seconds.

+++ spnl-core/src/main/scala/com/wajam/spnl/Task.scala
@@ -109,10 +109,10 @@ class Task(feeder: Feeder, val action: TaskAction, val persistence: TaskPersiste
     // [6,11], [7,12], [9,14], [13,18], [21,26], [37,42], [69,74], [133, 138], [261, 266]
     def updateUsingScatteredRandom {
       nextRetryTime =
-        lastErrorTime +                                                   // Initial timestamp offset
-        ExpectedUnavailableTimeInMs +                                     // Baseline Reboot Time
-        math.pow(2, retryCount).toLong * (1000 / context.throttleRate) +  // Exponentially increasing factor
-        (ExpectedUnavailableTimeInMs * util.Random.nextFloat()).toLong    // Random factor to scatter attempts
+        lastErrorTime +                                                    // Initial timestamp offset
+        ExpectedUnavailableTimeInMs +                                      // Baseline Reboot Time
+        (math.pow(2, retryCount) * (1000 / context.throttleRate) +         // Exponentially increasing factor
+        (ExpectedUnavailableTimeInMs * util.Random.nextFloat())).toLong    // Random factor to scatter attempts
     }
 
     // This formula is expected to be used in all other cases.
@@ -126,9 +126,9 @@ class Task(feeder: Feeder, val action: TaskAction, val persistence: TaskPersiste
     // [1, 2.5], [2, 3.5], [4, 5.5], [8, 9.5], [16, 17.5], [32, 33.5], [64, 65.5], [128, 129.5], [256, 257.5]
     def updateUsingConsistentRandom {
       nextRetryTime =
-        lastErrorTime +                                                   // initial timestamp offset
-        math.pow(2, retryCount).toLong * (1000 / context.throttleRate) +  // exponentially increasing factor
-        (1500 * util.Random.nextFloat()).toLong                           // slight random factor to spread traffic
+        lastErrorTime +                                              // initial timestamp offset
+        (math.pow(2, retryCount) * (1000 / context.throttleRate) +   // exponentially increasing factor
+        (1500 * util.Random.nextFloat())).toLong                     // slight random factor to spread traffic
     }
 
     def mustRetry = errorCount > retryCount
